Restructured external blockchain RPCs into blockchain package
