+++ blockchain/bitcoin/contracts/arc-bitcoin/bitcoin_read.go
@@ -8,7 +8,7 @@ import (
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/btcsuite/btcutil"
-	"github.com/republicprotocol/republic-go/bitcoin/client"
+	"github.com/republicprotocol/republic-go/blockchain/bitcoin"
 )
 
 type readResult struct {
