+++ blockchain/ethereum/bindings/generate.sh
